use printf instead of echo
Using "echo" in shell scripts is generally discouraged,
because most echo variants understand one or more options
but ignore the special "--" delimiter option.
They also frequently substitute some escaped characters.
This can lead to unexpected output
if the input starts with a dash or contains a backslash.

To circumvent all of those problems,
we now use the safer "printf %s ..." in its place.